maybe functional build tags?
